// front/src/ts/types/api.ts
export interface VideoFile {
  w: number;
  h: number;
  quality: string;
  codec?: string;
  quality_id?: number;
  file?: string;
  url?: {
    http: string;
    hls: string;
    hls2?: string;
    hls4?: string;
  };
  urls?: {
    http: string;
    hls: string;
    hls2?: string;
    hls4?: string;
  };
}

export interface AudioTrack {
  id: number;
  index: number;
  codec: string;
  channels: number;
  lang: string;
  type: { id: number; title: string; short_title: string | null };
  author: { id: number; title: string; short_title: string | null };
}

export interface Subtitle {
  lang: string;
  shift: number;
  embed: boolean;
  url: string;
}